requirements.txt
scipy==1.1.0
numpy==1.15.2
opencv_contrib_python==3.4.3.18
Pillow==6.1.0
PyQt5==5.13.0
sip==4.19.8